delivery_agent_service/requirements.txt
fastapi
uvicorn[standard]
httpx
tortoise-orm
asyncpg 
